+++ src/Client/ClientCore/Systems/ClientNetwork/ClientNetworkSystem.cs
@@ -45,6 +45,7 @@ namespace Sovereign.ClientCore.Systems.ClientNetwork
         {
             EventId.Client_Network_ConnectionLost,
             EventId.Client_Network_BeginConnection,
+            EventId.Client_Network_RegisterAccount,
         };
 
         public int WorkloadEstimate => 20;
@@ -59,9 +60,12 @@ namespace Sovereign.ClientCore.Systems.ClientNetwork
 
             eventDescriptions.RegisterNullEvent(EventId.Client_Network_Connected);
             eventDescriptions.RegisterNullEvent(EventId.Client_Network_ConnectionLost);
+            eventDescriptions.RegisterNullEvent(EventId.Client_Network_RegisterSuccess);
             eventDescriptions.RegisterEvent<BeginConnectionEventDetails>(EventId.Client_Network_BeginConnection);
             eventDescriptions.RegisterEvent<ErrorEventDetails>(EventId.Client_Network_ConnectionAttemptFailed);
             eventDescriptions.RegisterEvent<ErrorEventDetails>(EventId.Client_Network_LoginFailed);
+            eventDescriptions.RegisterEvent<RegisterAccountEventDetails>(EventId.Client_Network_RegisterAccount);
+            eventDescriptions.RegisterEvent<ErrorEventDetails>(EventId.Client_Network_RegisterFailed);
 
             eventLoop.RegisterSystem(this);
         }